[bug fixed] input and config classes
Config file changed: now the user can specify the time unit, the figure width and figure height.

Input changed: fixed a bug that affected the file size calculations (overflow of integer prevented the user
to open large files)

+++ common/input.hpp
@@ -32,7 +32,7 @@ public:
     char plainLength;
 private:
     FILE* file;
-    long fileSize;
+    long long fileSize;
     int getTraceOffset(int traceIndex);
     int getPlainOffset(int plainIndex);
 };
